Remove the now-unnecessary new ScripCondition
Splitting the condition logic for this into a ScripCondition does two
things: allows reuse of the condition, and potentially allows the
action to be called with a different condition.  The latter is
entirely confusing, and the former is odd and orthogonal to the
purpose of this extension.

As the new ScripAction only makes sense with that one condition, and
as of the previous commit it not encodes the entirely of its necessary
logic in `Prepare`, remove the Condition.